Redoing to proofs in SetTheory2 (pairs symmetry and deconstruction). Making some improvements to applySubst
