+++ lisa-utils/src/main/scala/lisa/prooflib/SimpleDeducedSteps.scala
@@ -304,4 +304,8 @@ object SimpleDeducedSteps {
       }
     }
   }
+
+
+
+  object ByEquivalence
 }
