+++ src/main/scala/lisa/proven/mathematics/SetTheoryTactics.scala
@@ -11,26 +11,13 @@ import lisa.proven.mathematics.SetTheory2
 import lisa.utils.KernelHelpers.*
 import lisa.utils.Printer
 
-object SetTheoryTactics extends lisa.Main {
+object SetTheoryTactics  {
 
+  import lisa.settheory.SetTheoryLibrary.{_, given}
   // var defs
-  private val w = variable
   private val x = variable
   private val y = variable
   private val z = variable
-  private val t = variable
-  private val a = variable
-  private val b = variable
-  private val c = variable
-  private val d = variable
-
-  // relation and function symbols
-  private val r = variable
-  private val f = variable
-  private val g = variable
-
-  private val P = predicate(1)
-  private val Q = predicate(1)
   private val schemPred = predicate(1)
 
   /**
@@ -56,7 +43,7 @@ object SetTheoryTactics extends lisa.Main {
    * See [[setIntersection]] or [[relationDomain]] for more usage.
    */
   object UniqueComprehension extends ProofTactic {
-    def apply(using proof: Library#Proof, line: sourcecode.Line, file: sourcecode.File)(originalSet: Term, separationPredicate: LambdaTermFormula)(bot: Sequent): proof.ProofTacticJudgement = {
+    def apply(using proof: Library#Proof, line: sourcecode.Line, file: sourcecode.File, om: OutputManager)(originalSet: Term, separationPredicate: LambdaTermFormula)(bot: Sequent): proof.ProofTacticJudgement = {
       require(separationPredicate.vars.length == 2) // separationPredicate takes two args
 
       // fresh variable names to avoid conflicts
