--- src/test/scala/lisa/examples/ExampleTests.scala
@@ -1,29 +1,0 @@
-package lisa.examples
-
-import org.scalatest.funsuite.AnyFunSuite
-
-import scala.language.adhocExtensions
-
-class ExampleTests extends AnyFunSuite {
-  /*
-  test("front interactive proof (1)") {
-    frontInteractiveProof1()
-  }
-
-  test("front interactive proof (2)") {
-    frontInteractiveProof2()
-  }
-
-  test("front matching") {
-    frontMatching()
-  }
-
-  test("front parsing & printing") {
-    frontParsingPrinting()
-  }
-
-  test("front solver") {
-    frontSolver()
-  }
-   */
-}
