+++ src/test/scala/lisa/examples/peano_example/Peano.scala
@@ -1,4 +1,4 @@
-package lisa.proven.peano_example
+package lisa.examples.peano_example
 
 import lisa.automation.kernel.ProofTactics.*
 import lisa.kernel.fol.FOL.*
