+++ src/test/scala/lisa/examples/peano_example/PeanoArithmeticsLibrary.scala
@@ -1,0 +1,7 @@
+package lisa.examples.peano_example
+
+import lisa.examples.peano_example
+
+object PeanoArithmeticsLibrary extends lisa.prooflib.Library(peano_example.PeanoArithmetics.runningPeanoTheory) {
+  export PeanoArithmetics.*
+}
